Danbooru search improve what images can be senr
Every time Telegram gives us an error that it can't send a group, we just send the next image as normal photo. Like this we remove the self checking time and can ensure that as many images as possible are sendt.
There was also the bug with the error message: "Group send failed" which I don't know why I got it. This has been resolved too now. All images which can not be sent as normal photos are now added to the error list instead of all when getting this error.